PageCache header created.
